update lowongan pekerjaan

+++ database/migrations/2026_03_27_075522_add_foto_to_tb_alokasi_punia_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('tb_alokasi_punia', function (Blueprint $table) {
-            $table->json('foto')->nullable();
+            if (!Schema::hasColumn('tb_alokasi_punia', 'foto')) {
+                $table->json('foto')->nullable();
+            }
         });
     }
 
@@ -22,7 +24,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('tb_alokasi_punia', function (Blueprint $table) {
-            $table->dropColumn('foto');
+            if (Schema::hasColumn('tb_alokasi_punia', 'foto')) {
+                $table->dropColumn('foto');
+            }
         });
     }
 };
